move enable() method out of context
